trying to fix stuffs
