--- Compiled/CSS/block.css
@@ -1,3 +1,0 @@
-
-
-/*# sourceMappingURL=block.css.map */
