--- Compiled/CSS/composition.css
@@ -1,3 +1,0 @@
-
-
-/*# sourceMappingURL=composition.css.map */
