--- Compiled/JS/Models/Reservations.js
@@ -1,2 +1,0 @@
-export class Reservations {
-}
