--- Compiled/JS/Models/Table.js
@@ -1,2 +1,0 @@
-export class Table {
-}
