+++ register.html
@@ -4,7 +4,6 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>הרשמה</title>
-    <script defer src="../views/view.js"></script>
 </head>
 <body>
     <h1>הרשמה</h1>
@@ -19,5 +18,6 @@
         <button type="submit">submit</button>
     </form>
     <p>Already have account? <a href="login.html">Login</a></p>
+    <script type="module" src="dist/view.js"></script>
 </body>
 </html>
